Move SummarizeTestResult to its own distbench_summary.cc file
